requirements.txt
absl-py==0.7.1
astor==0.7.1
astroid==2.2.5
audioread==2.1.6
cycler==0.10.0
decorator==4.4.0
Django==2.2
djangorestframework==3.9.2
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
imageio==2.5.0
isort==4.3.16
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
librosa==0.6.3
llvmlite==0.28.0
Markdown==3.1
matplotlib==3.0.3
mccabe==0.6.1
mock==2.0.0
networkx==2.2
numba==0.43.1
numpy==1.16.2
pbr==5.1.3
Pillow==6.0.0
protobuf==3.7.1
pylint==2.3.1
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2018.9
PyWavelets==1.0.2
PyYAML==5.1
resampy==0.2.1
scikit-image==0.15.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
sklearn==0.0
sqlparse==0.3.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
typed-ast==1.3.1
Werkzeug==0.15.2
wrapt==1.11.1
